plumbing: format, Fix Packfile to return rbuf only when available
When Packfile.Close() gets called while it is not initialized(p.init() isn't called),
nil reader(p.rbuf) is passed to the pool. This results in panic since following
calls to sync.GetBufioReader() would return the nil reader. This can happen if
the user only invokes methods of idxfile.Index, as it is embeded to the packfile.
Now Packfile.Close() checks if p.rbuf is nil before returning it  to the pool.

The issue was exposed by tests in 'tests/pack/pack_test.go'. Since it is now fixed,
resetGlobalSyncPools() function also removed.